rozděleno do dvou souborů. Řešení téhož úkolu, jednou s využitím funkce, jednou bez
